feat: add COOKIE_SECURE (default: true) and COOKIE_SAME_SITE (default: lax) environment variables
fix #293
