+++ docs/configuration.md
@@ -436,6 +436,8 @@ This gets passed into the `employedBy()` function, (called `companyId`), and can
 | `ALLOWED_EMAIL_DOMAINS`      |                         | List of comma-separated email domain names that are allowed to register.                                                                                                                     |
 | `AUTO_ACTIVATE_NEW_USERS`    | false                   | When set to true, automatically activate the users once registererd.                                                                                                                         |
 | `COOKIE_SECRET`              |                         |                                                                                                                                                                                              |
+| `COOKIE_SECURE`              | false                   | Secure cookie attribute                                                                                                                                                                      |
+| `COOKIE_SAME_SITE`           | lax                     | SameSite cookie attribute. Accepts true, false, lax, none or strict. If none of these values it will be set to 'lax'                                                                         |
 | `DEFAULT_USER_ROLE`          | user                    |                                                                                                                                                                                              |
 | `DEFAULT_ALLOWED_USER_ROLES` | user                    | Comma spearated list of allowed roles assigned to each user. Defaults to DEFAULT_USER_ROLE.                                                                                                  |
 | `HIBP_ENABLE`                | false                   |                                                                                                                                                                                              |
