+++ src/routes/auth/login.ts
@@ -3,7 +3,8 @@ import Boom from '@hapi/boom'
 import bcrypt from 'bcryptjs'
 import { v4 as uuidv4 } from 'uuid'
 import { asyncWrapper, selectAccount } from '@shared/helpers'
-import { newJwtExpiry, setRefreshToken, createHasuraJwt } from '@shared/jwt'
+import { newJwtExpiry, createHasuraJwt } from '@shared/jwt'
+import { setRefreshToken } from '@shared/cookies'
 import { loginAnonymouslySchema, loginSchema } from '@shared/validation'
 import { insertAccount } from '@shared/queries'
 import { request } from '@shared/request'
